Dataset-Construction/requirements.txt
requests==2.32.3
pillow==11.1.0